completed images table

+++ database/migrations/2022_11_27_112039_create_images.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('images', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger("productsid");
+            $table->integer("productsid");
             $table->string("Name");
             $table->string("Image1");
             $table->string("Image2");
